CB-16360 Move AccountAwareResource and AccountIdAwareResource and related classes into the common. Most of the entities have resourceCrn, name and id. Before I integrate the queries into the AccountAwareService, I move into the common module
