+++ core-model/src/test/java/com/sequenceiq/cloudbreak/domain/EntityTest.java
@@ -11,7 +11,7 @@ import org.junit.Test;
 import org.reflections.Reflections;
 
 import com.sequenceiq.cloudbreak.service.secret.SecretValue;
-import com.sequenceiq.cloudbreak.service.secret.domain.AccountIdAwareResource;
+import com.sequenceiq.cloudbreak.common.dal.model.AccountIdAwareResource;
 import com.sequenceiq.cloudbreak.workspace.model.TenantAwareResource;
 
 public class EntityTest {
