+++ environment/src/main/java/com/sequenceiq/environment/environment/service/EnvironmentService.java
@@ -39,8 +39,8 @@ import com.sequenceiq.cloudbreak.common.service.TransactionService;
 import com.sequenceiq.cloudbreak.logger.MDCBuilder;
 import com.sequenceiq.cloudbreak.logger.MDCUtils;
 import com.sequenceiq.cloudbreak.quartz.model.JobResource;
-import com.sequenceiq.cloudbreak.structuredevent.repository.AccountAwareResourceRepository;
-import com.sequenceiq.cloudbreak.structuredevent.service.AbstractAccountAwareResourceService;
+import com.sequenceiq.cloudbreak.common.dal.repository.AccountAwareResourceRepository;
+import com.sequenceiq.cloudbreak.common.service.account.AbstractAccountAwareResourceService;
 import com.sequenceiq.environment.environment.EnvironmentStatus;
 import com.sequenceiq.environment.environment.domain.Environment;
 import com.sequenceiq.environment.environment.domain.Region;
