+++ environment/src/main/java/com/sequenceiq/environment/parameters/dao/domain/AwsParameters.java
@@ -3,7 +3,7 @@ package com.sequenceiq.environment.parameters.dao.domain;
 import static com.sequenceiq.cloudbreak.util.NullUtil.getIfNotNull;
 
 import com.sequenceiq.cloudbreak.service.secret.SecretValue;
-import com.sequenceiq.cloudbreak.service.secret.domain.AccountIdAwareResource;
+import com.sequenceiq.cloudbreak.common.dal.model.AccountIdAwareResource;
 import com.sequenceiq.cloudbreak.service.secret.domain.Secret;
 import com.sequenceiq.cloudbreak.service.secret.domain.SecretToString;
 import com.sequenceiq.environment.parameter.dto.s3guard.S3GuardTableCreation;
