+++ freeipa/src/main/java/com/sequenceiq/freeipa/entity/UserSyncStatus.java
@@ -11,7 +11,7 @@ import javax.persistence.SequenceGenerator;
 
 import com.sequenceiq.cloudbreak.common.json.Json;
 import com.sequenceiq.cloudbreak.common.json.JsonToString;
-import com.sequenceiq.cloudbreak.service.secret.domain.AccountIdAwareResource;
+import com.sequenceiq.cloudbreak.common.dal.model.AccountIdAwareResource;
 
 @Entity
 public class UserSyncStatus implements AccountIdAwareResource {
