+++ freeipa/src/main/java/com/sequenceiq/freeipa/ldap/LdapCustomCrnOrNameProvider.java
@@ -6,7 +6,7 @@ import javax.inject.Inject;
 
 import org.springframework.stereotype.Component;
 
-import com.sequenceiq.cloudbreak.structuredevent.repository.AccountAwareResource;
+import com.sequenceiq.cloudbreak.common.dal.model.AccountAwareResource;
 import com.sequenceiq.freeipa.events.AbstractCustomCrnOrNameProvider;
 
 @Component
