+++ secret-engine/src/main/java/com/sequenceiq/cloudbreak/service/secret/service/SecretAspectService.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Service;
 import com.google.common.base.Strings;
 import com.sequenceiq.cloudbreak.service.secret.SecretOperationException;
 import com.sequenceiq.cloudbreak.service.secret.SecretValue;
-import com.sequenceiq.cloudbreak.service.secret.domain.AccountIdAwareResource;
+import com.sequenceiq.cloudbreak.common.dal.model.AccountIdAwareResource;
 import com.sequenceiq.cloudbreak.service.secret.domain.Secret;
 import com.sequenceiq.cloudbreak.service.secret.domain.SecretProxy;
