+++ structuredevent-service-cdp/src/main/java/com/sequenceiq/cloudbreak/structuredevent/domain/CDPStructuredEventEntity.java
@@ -13,7 +13,7 @@ import com.sequenceiq.cloudbreak.common.json.Json;
 import com.sequenceiq.cloudbreak.common.json.JsonToString;
 import com.sequenceiq.cloudbreak.structuredevent.service.converter.CDPStructuredEventTypeConverter;
 import com.sequenceiq.cloudbreak.structuredevent.event.StructuredEventType;
-import com.sequenceiq.cloudbreak.structuredevent.repository.AccountAwareResource;
+import com.sequenceiq.cloudbreak.common.dal.model.AccountAwareResource;
 
 @Entity
 @Table(name = "cdp_structured_event")
